requirements.txt
fastapi 
pymongo 
uvicorn
python-dotenv
python-multipart
pandas
openai
langchain_openai 
langchain_text_splitters 
faiss-cpu 
numpy
groq
langchain-groq
google-genai
google-generativeai
agno
langchain
langgraph
langchain_community
streamlit
requests
oauth2client
retrying
graphviz
huggingface_hub